+++ test/WorkerExtensionTests/EventHubs/EventDataConverterTests.cs
@@ -67,6 +67,104 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
             AssertEventData(output[1]);
         }
 
+        [Fact]
+        public async Task ConvertAsync_ReturnsFailure_WrongContentType()
+        {
+            var eventData = CreateEventData();
+
+            var data = new GrpcModelBindingData(new ModelBindingData()
+            {
+                Version = "1.0",
+                Source = Constants.BindingSource,
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(eventData)),
+                ContentType = "application/json"
+            });
+            var context = new TestConverterContext(typeof(string), data);
+            var converter = new EventDataConverter();
+            var result = await converter.ConvertAsync(context);
+
+            Assert.Equal(ConversionStatus.Failed, result.Status);
+            var output = result.Value as EventData;
+            Assert.Null(output);
+            Assert.IsType<InvalidOperationException>(result.Error);
+        }
+
+        [Fact]
+        public async Task ConvertAsync_Batch_ReturnsFailure_WrongContentType()
+        {
+            var message = CreateEventData();
+
+            var data = new ModelBindingData
+            {
+                Version = "1.0",
+                Source = Constants.BindingSource,
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(message)),
+                ContentType = "application/json"
+            };
+
+            var array = new CollectionModelBindingData();
+            array.ModelBindingData.Add(data);
+            array.ModelBindingData.Add(data);
+
+            var context = new TestConverterContext(typeof(string), new GrpcCollectionModelBindingData(array));
+            var converter = new EventDataConverter();
+            var result = await converter.ConvertAsync(context);
+
+            Assert.Equal(ConversionStatus.Failed, result.Status);
+            var output = result.Value as EventData[];
+            Assert.Null(output);
+            Assert.IsType<InvalidOperationException>(result.Error);
+        }
+
+        [Fact]
+        public async Task ConvertAsync_ReturnsFailure_WrongSource()
+        {
+            var eventData = CreateEventData();
+
+            var data = new GrpcModelBindingData(new ModelBindingData()
+            {
+                Version = "1.0",
+                Source = "some-other-source",
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(eventData)),
+                ContentType = Constants.BinaryContentType
+            });
+            var context = new TestConverterContext(typeof(string), data);
+            var converter = new EventDataConverter();
+            var result = await converter.ConvertAsync(context);
+
+            Assert.Equal(ConversionStatus.Failed, result.Status);
+            var output = result.Value as EventData;
+            Assert.Null(output);
+            Assert.IsType<InvalidOperationException>(result.Error);
+        }
+
+        [Fact]
+        public async Task ConvertAsync_Batch_ReturnsFailure_WrongSource()
+        {
+            var message = CreateEventData();
+
+            var data = new ModelBindingData
+            {
+                Version = "1.0",
+                Source = "some-other-source",
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(message)),
+                ContentType = Constants.BinaryContentType
+            };
+
+            var array = new CollectionModelBindingData();
+            array.ModelBindingData.Add(data);
+            array.ModelBindingData.Add(data);
+
+            var context = new TestConverterContext(typeof(string), new GrpcCollectionModelBindingData(array));
+            var converter = new EventDataConverter();
+            var result = await converter.ConvertAsync(context);
+
+            Assert.Equal(ConversionStatus.Failed, result.Status);
+            var output = result.Value as EventData[];
+            Assert.Null(output);
+            Assert.IsType<InvalidOperationException>(result.Error);
+        }
+
         private static void AssertEventData(EventData output)
         {
             Assert.Equal("body", output.EventBody.ToString());