+++ extensions/Worker.Extensions.ServiceBus/release_notes.md
@@ -7,3 +7,4 @@
 ### Microsoft.Azure.Functions.Worker.Extensions.ServiceBus 5.10.0-preview3
 
 - Make Constants class internal (#1609)
+- Return ConversionResult.Failed when the wrong content-type or binding source is used (#1609)